requirements.txt
facebook-sdk==3.1.0
flask==1.1.1
requests==2.22.0
psycopg2-binary==2.8.4